Bug fix - Camera transitions sometimes incorrect
In conditions where there are multiple overlapping camera position
triggers, there was a bug that chooses the last entered trigger as the
camera position. Sometimes this is wrong, added a list of camera
positions we are currently triggering, on triggerExit removing the
camera position we are leaving, and setting a new one to the last one in
the list.